Ticket #22 Saber Block - Valkyra [fixed] compiler warnings
git-svn-id: http://svn2.assembla.com/svn/swgemu/core3/trunk/MMOCoreORB@1329 c3d1530f-68f5-4bd0-87dc-8ef779617e40

+++ src/server/zone/objects/creature/CreatureObjectImplementation.h
@@ -1102,7 +1102,7 @@ public:
 	}
 
 	inline int getConditionDamage() {
-		return conditionDamage;
+		return (int)conditionDamage;
 	}
 
 	inline int getMaxCondition() {
@@ -1110,7 +1110,7 @@ public:
 	}
 
 	inline int getCondition() {
-		return maxCondition - conditionDamage;
+		return (int)(maxCondition - conditionDamage);
 	}
 
 	inline int getCenteredBonus() {
